+++ QUANTAXIS/__init__.py
@@ -31,7 +31,7 @@ by yutiansut
 2017/4/8
 """
 
-__version__ = '1.5.30'
+__version__ = '1.6.0'
 __author__ = 'yutiansut'
 
 import argparse
