+++ setup.py
@@ -75,12 +75,13 @@ PACKAGES = ["QUANTAXIS", "QUANTAXIS.QAFetch", "QUANTAXIS.QACmd", "QUANTAXIS.QAMa
 DESCRIPTION = "QUANTAXIS:Quantitative Financial Strategy Framework"
 
 
-try:
-    import pypandoc
-    LONG_DESCRIPTION = pypandoc.convert_file('README.md', 'rst')
-except Exception:
-    with open("README_ENG.md", "r", encoding='utf-8') as fh:
-        LONG_DESCRIPTION = fh.read()
+# try:
+#     import pypandoc
+#     LONG_DESCRIPTION = pypandoc.convert_file('README.md', 'rst')
+# except Exception:
+# with open("README_ENG.md", "r", encoding='utf-8') as fh:
+#     LONG_DESCRIPTION = fh.read()
+LONG_DESCRIPTION ='QUANTAXIS Financial Framework'
 
 """
